requirements.txt
pyrofork
tgcrypto
requests
aiohttp
telethon
asyncio
pytz
telegraph
pycountry
beautifulsoup4
pillow
yt-dlp
spotipy
python-dateutil
pymongo
gTTS
aiofiles
python-dotenv
googletrans==4.0.0-rc1
psutil
pydub
moviepy
speedtest-cli
